% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_neighbours.R
\name{add_neighbours}
\alias{add_neighbours}
\title{Adds adjacent neighbours to transect data}
\usage{
add_neighbours(dat_pts, template)
}
\arguments{
\item{dat_pts}{training points/pixels from sampled transects}

\item{template}{the baseline raster template for the entire map area}
}
\value{
an sfpoints object
}
\description{
For use in the accuracy metrics this function adds all surrounding pixels to each origin pixel in the training data
}
\examples{
tpoints_ne <- add_neighbours(tpoints,trast)
}
\keyword{accuracy}
\keyword{assessment,}
\keyword{data,}
\keyword{neighbours}
\keyword{training}
