% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_points.R
\name{attibute_points}
\alias{attibute_points}
\title{Add covariate predictor data to training point file}
\usage{
attibute_points(training, cov_dir, res_dir)
}
\arguments{
\item{training}{a file gpkg of training points}

\item{cov_dir}{folder containing covariate files}

\item{res_dir}{resolution subfolder}

\item{output}{Location of where attributed training file will be exported}
}
\description{
Generates base landscape covariates from a 25m TRIM DEM using SAGA GIS and converts to classes for use in the sample plan cLHS
}
\details{
This script has been tested with SAGA 8.4 on Windows
Depending on your system the path to \code{saga_cmd} may need to be specified.
}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
