% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_pts.R
\name{attribute_points}
\alias{attribute_points}
\title{Add covariate predictor data to training point file}
\usage{
attribute_points(
  data_pts,
  cov_dir,
  write_output = TRUE,
  out_dir = fs::path(PEMprepr::read_fid()$dir_20105020_clean_field_data$path_rel,
    out_name = "allpoints.gpkg"),
  out_name = "allpoints_att.gpkg"
)
}
\arguments{
\item{cov_dir}{folder containing covariate files}

\item{write_output}{A \code{logical}if the sf object be written to disk?
If \code{TRUE} (default), will write to \code{out_dir} under the appropriate resolution subfolder.}

\item{out_dir}{A character string of path which points to output location. A default
location and name are applied in line with standard workflow.}

\item{out_name}{A character string of the output file name. Default is \code{allpoints_att.gpkg}}

\item{dat_pts}{a file gpkg of training points}
}
\value{
an sf object
}
\description{
Add covariate predictor data to training point file
}
\examples{
\dontrun{
tpoints_ne <- attribute_points(dat_pts, cov_dir)
}
}
