% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sampleplan.R
\name{build_sampleplan}
\alias{build_sampleplan}
\title{build_sampleplan}
\usage{
build_sampleplan(
  boi,
  clhs_set,
  sampleplan_input,
  basedata_input,
  output,
  outname = "s1_sampling"
)
}
\arguments{
\item{boi}{string with name of bgc varies ie: "ICHmc2"}

\item{clhs_set}{sring name of repreat clhs output to use in final sample plan, ie: "ICHmc2_clhs_sample_3.gpkg"}

\item{sampleplan_input}{filepath where repeat clhs outputs are saved}

\item{basedata_input}{filepath of base sample 1 inputs, root folder for landscape covariates and exclusion folders}

\item{output}{filepath where gpkg with transects are saved}

\item{outname}{string output file name, default is "s1_sampling"}
}
\value{
sf final sampleplan
}
\description{
build_sampleplan
}
\examples{
build_sampleplan("ICHmc2", "ICHmc2_clhs_sample_3.gpkg", sampleplan_input = fid$samplingplan_clhs[2],
                basedata_input = fid$sampling_input_2010[2]
                out_path <- fid$sampleplan_final_transect[2]),outname = "s1_sampling")
}
