% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bgc_landscape.R
\name{check_bgc_landscapes}
\alias{check_bgc_landscapes}
\title{Generate a summary table of binned landscape classes per BGC}
\usage{
check_bgc_landscapes(
  bec = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel, "25m", "bec.tif"),
  binned_landscape = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel, "25m",
    "modules_landscape", "landscape_binned.tif")
)
}
\arguments{
\item{bec}{A \code{SpatRast} of path to BEC raster generated at a landscape (25m)
scale. This output is derived from the create_bgc_template().}

\item{binned_landscape}{A \code{SpatRast} with the landscape binned. This is the
output of the \code{\link[=create_binned_landscape]{create_binned_landscape()}} function.}
}
\value{
A dataframe with summary of landsclass class and BEC unit.
A plot is also returned.
}
\description{
Generate a summary table of binned landscape classes per BGC
}
\examples{
\dontrun{
check_bgc_landscapes = create_bgc_template(
  bec = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel,
  "25m", "bec.tif")
  binned_landscape = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel,
  "25m","modules_landscape")
}
}
