% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locations.R
\name{check_locations}
\alias{check_locations}
\title{Check start locations meet sample plan criteria}
\usage{
check_locations(roads, locations)
}
\arguments{
\item{roads}{An \code{sf} object or path to roads layers. Default location is based
on standard workflow. Roads is created using the create_base_vectors()
function.}

\item{locations}{An \code{sf} object with start locations points. These represent the
start locations (such as a town or location) from which the cost layer
will be based. Locations may be single or multi point.}
}
\value{
TRUE
}
\description{
This functions checks that selected start points are all or partially
connected to the road network
}
\examples{
\dontrun{
roadsls <- get_roads(aoils, PEMprepr::read_fid()$dir_201010_inputs$path_abs)
cities <- sf::st_read(fs::path(PEMprepr::read_fid()$dir_1010_vector$path_rel,"major_towns_bc.gpkg"))
nearest_town = "Hazelton"
start <- cities[cities$NAME == nearest_town,"NAME"]
check_locations(roadsls, start)
}
}
