% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_road_layer.R
\name{check_road_layer}
\alias{check_road_layer}
\title{Check major road layer
Runs a number of basic checks on the road layer prior to using in costlayer}
\usage{
check_road_layer(roads)
}
\arguments{
\item{roads}{sf object "road_major.gpkg"}
}
\value{
TRUE if all checks run correctly
}
\description{
Check major road layer
Runs a number of basic checks on the road layer prior to using in costlayer
}
\examples{
check_road_layer(roads)
}
