% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_roads.R
\name{check_roads}
\alias{check_roads}
\title{Check road network layer}
\usage{
check_roads(
  roads = fs::path(PEMprepr::read_fid()$dir_1010_vector$path_rel, "road_network.gpkg")
)
}
\arguments{
\item{roads}{a \code{sf} object or path to roads layers. Default location is based
on standard workflow. Roads is created using the create_base_vectors()
function.}
}
\value{
TRUE
}
\description{
Check road network layer
}
\examples{
\dontrun{
roads_raw <- sf::st_read(
fs::path(PEMprepr::read_fid()$dir_1010_vector$path_rel,"road_network.gpkg"))
check_road_layer(roads_raw)
}
}
