% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_bgc_cost.R
\name{classify_bgc_cost}
\alias{classify_bgc_cost}
\title{Classify the sample cost per BGC variant and subzone}
\usage{
classify_bgc_cost(bec, binned_landscape, cost)
}
\arguments{
\item{bec}{A \code{SpatRast} or path to BEC raster generated at a landscape (25m)
scale. This output is derived from the create_bgc_template().}

\item{binned_landscape}{A \code{SpatRast} with the landscape binned. This is the
output of the \code{\link[=create_binned_landscape]{create_binned_landscape()}} function.}

\item{cost}{A \code{SpatRast} or path to the generated cost layer. This is the
output of the function (prep_cost_layer TBFinalised)}
}
\value{
data.frame with a class
}
\description{
Assess the costs of sampling areas for each BGC to help assess how well the cost layer describes to study area
}
\examples{
\dontrun{
bec <- terra::rast(
  fs::path(PEMprepr::read_fid()$dir_201010_inputs$path_abs, "25m", "bec.tif"))
binned_landscape <- terra::rast(
    fs::path(PEMprepr::read_fid()$dir_201010_inputs$path_abs, "25m",
    "modules_landscape", "landscape_binned.tif"))
cost <- terra::rast(
    fs::path(PEMprepr::read_fid()$dir_201010_inputs$path_abs,"acost.tif"))
check_bgc_cost(bec, binned_landscape, cost)
}
}
