% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aspect_classes.R
\name{create_aspect_classes}
\alias{create_aspect_classes}
\title{Create aspect classes}
\usage{
create_aspect_classes(x, threshold = 0.2)
}
\arguments{
\item{x:}{input 25m dah created in SAGA (.tif).}

\item{threshold:}{sets the DAH threshold where <x are cold aspects and >x are warm aspects. The rest are neutral.}
}
\description{
Creates a 3-class aspect from DAH. classes from a TPI generated from a DEM following the rules in
https://github.com/gianmarcoalberti/GmAMisc/blob.master/R/landfClass.r
}
\details{
Thresholds of 0.2 - 0.3 are reasonable setting
}
\examples{
data(dah)

}
\keyword{anisotropic}
\keyword{aspect}
\keyword{classes}
\keyword{diurnal}
\keyword{heating,}
