% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clhs.R
\name{create_clhs}
\alias{create_clhs}
\title{Create CLHS sample plan
This function takes a terra stack of covariates (and cost), and uses
conditional latin hypercube sampling to create the specified number of points
in slices.}
\usage{
create_clhs(
  all_cov,
  num_slices,
  to_include = NULL,
  n_points = 5,
  min_dist = 900,
  num_sample = 5e+06
)
}
\arguments{
\item{all_cov}{A \code{spatRast} stack with all landscape covariates and masked
cost layer. Note all layers excluding the cost layer should be factors.}

\item{num_slices}{A \code{numeric} representing the number of slices to be built}

\item{to_include}{A \code{sf}dataframe of already sample points to include in plan}

\item{n_points}{A \code{numeric} represrnting the numnber of points per slice}

\item{min_dist}{A `numeric' value for the minimum distance to between each
clhs point in meters. Default is 1000m.}

\item{num_sample}{A \code{numeric}for the number of samples to run CLHS on.
Default is 5000000.}
}
\value{
\code{sf} points object with the location of clhs sample points
}
\description{
Create CLHS sample plan
This function takes a terra stack of covariates (and cost), and uses
conditional latin hypercube sampling to create the specified number of points
in slices.
}
\examples{
\dontrun{
create_clhs(all_cov = sample_layers_masked,
                num_slices = 3,
}
}
