% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_exclusion.R
\name{create_cost_exclusion}
\alias{create_cost_exclusion}
\title{Add exclusion areas to the cost layer for sampling}
\usage{
create_cost_exclusion(
  vec_dir = fs::path(PEMprepr::read_fid()$dir_1010_vector$path_abs),
  cost,
  buffer = 150,
  out_dir = fs::path(PEMprepr::read_fid()$dir_201010_inputs$path_abs),
  write_output = TRUE
)
}
\arguments{
\item{vec_dir}{A \code{character} string or path which points to base vector
layers created using create_base_vectors(). A default location and name
are applied in line with standard workflow.}

\item{cost}{A \code{SpatRast} or path to cost layer with high cost applied.
This is created using the create_cost_penality() function. A default
location and name is applied in line with standard workflow.}

\item{buffer}{A \code{numeric} values in meters representing the buffer distance to
be excluded surrounding lakes}

\item{out_dir}{A \code{character} or path which points to input location
of . A default location and name are applied in line with standard workflow.}

\item{write_output}{A \code{logical} should the cost_penalty spatRaster be
written to disk? If \code{TRUE} (default), will write to \code{out_dir}.}
}
\value{
A \code{SpatRast}**` A masked cost raster layer
}
\description{
Creates and applies a mask to permanently changed areas where sample is not be conducted.
This includes (lakes, and permanently changed landscape i.e roads)
}
\examples{
\dontrun{
cost_masked <- create_cost_nosample(vec_dir = vec_dir,
                                   cost = cost_penalty,
                                 buffer = 150,
                                 write_output = FALSE)

}

}
