% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_nosample.R
\name{create_cost_nosample}
\alias{create_cost_nosample}
\title{Create cost no sample area}
\usage{
create_cost_nosample(vec_dir, cost)
}
\arguments{
\item{vec_dir}{text string with folder location of base vector layers}

\item{cost}{SpatRast of cost layer with high cost penalty applied}
}
\value{
\strong{SpatRast} A masked cost raster layer
}
\description{
Creates and applies a mask to permanently changed areas where sample is not be conducted.
This includes (lakes, and permanently changed landscape i.e roads)
}
\examples{
create_cost_nosample <- function(vec_dir, cost)

}
