% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_helicost_layer.R
\name{create_helicost_layer}
\alias{create_helicost_layer}
\title{create cost layer}
\usage{
create_helicost_layer(dem, start)
}
\arguments{
\item{dem}{**SpatRast or RasterLayer (raster package), output from prepare_cost_layer ()}

\item{start}{**SpatVec of sf object with Start Locations for the cost layer.}
}
\value{
RasterLayer with Cost
}
\description{
create cost layer
}
\examples{
create_cost(costprep, start)
}
