% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_landform_classes.R
\name{create_landform_classes}
\alias{create_landform_classes}
\title{Create landform classes}
\usage{
create_landform_classes(
  x,
  scale = 3,
  sn = 3,
  ln = 7,
  n.classes = "six",
  add.tpi = FALSE,
  stand.tpi = FALSE
)
}
\arguments{
\item{x:}{input 25m dem (.tif).}

\item{scale:}{size (in terms of cells per side) of the neighborhood (moving window) to be used; it must be an odd integer.}

\item{sn:}{if the 10-class classification is selected, this paramenter sets the s(mall) n(eighborhood) to be used.}

\item{ln:}{if the 10-class classification is selected, this paramenter sets the l(arge) n(eighborhood) to be used.}

\item{n.classes:}{"six" or "ten" for a six- or ten-class landform classification.}

\item{add.tpi:}{set to TRUE will return a TPI raster (FALSE is default).}

\item{stand.tpi:}{specifies whether the returned TPI raster will be un- or standardized (FALSE is default).}
}
\description{
Generates a TPI layer from a 25m DEM standard and  converts to landform classes from a following the rules in
https://github.com/gianmarcoalberti/GmAMisc/blob/master/R/landfClass.R
}
\details{
Classes derived from TPI calculated from a 25m DEM
}
\examples{
data(dem)

}
\keyword{classes}
\keyword{landform,}
