% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_binned_landscape.R
\name{create_landscape_bins}
\alias{create_landscape_bins}
\title{Create binned landscape}
\usage{
create_landscape_bins(input)
}
\arguments{
\item{input}{text string with either folder location or individual files}
}
\value{
\strong{SpatRaster}
}
\description{
Create a landscape with binned variables to assess the landscape environmental space.
}
\details{
Can use either folder or individual files to assess landscape covariate space
}
\examples{

create_landscape_bins()
}
