% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vrp.R
\name{create_vrp}
\alias{create_vrp}
\title{Vehicle Routing Problem}
\usage{
create_vrp(x, clhs_points, start_pos, number_days)
}
\arguments{
\item{x}{leastcostpath conductance matrix}

\item{clhs_points}{sample locations (sf)}

\item{start_pos}{start position (sf)}

\item{number_days}{number of days to sample for}
}
\value{
list of vrp details
}
\description{
Vehicle Routing Problem
}
\author{
Kiri Daust
}
