% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site_transects.R
\name{.rot}
\alias{.rot}
\title{Build site transects
Build the associates triangles, buffered triangles and selected paired triangle to sample, based on lowest cost.}
\usage{
.rot(a)
}
\arguments{
\item{sample_points}{\strong{sf} spatial object of clhs points}

\item{cost}{\strong{SpatRast} cost layer generate for sample plan}

\item{centroid_distance}{Numeric value at which the triangles are placed apart, default is 400 based on albers crs meters}

\item{out_path}{text string with location in which output sample plan as a geopackage is written}
}
\value{
TRUE
}
\description{
Build site transects
Build the associates triangles, buffered triangles and selected paired triangle to sample, based on lowest cost.
}
\examples{
build_site_transects(sample_points, cost, centroid_distance = 400, out_path)
}
