% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fielddata.R
\name{format_fielddata}
\alias{format_fielddata}
\title{Format field data}
\usage{
format_fielddata(data_dir = NULL, transect_layout, buffer = 10)
}
\arguments{
\item{data_dir}{text string with location of raw files in shp or gpk format}

\item{transect_layout}{sf object with simplified transect layout}

\item{buffer}{numeric value for buffer distance around transect layout}
}
\value{
sf point data with standardized and consolidated datasets
}
\description{
Standardizes format from raw field data to standard point data
}
\examples{
\dontrun{
clean_pts <- format_fielddata(inputfolder, transect_layout, buffer = 10)
}
}
