% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tracklog.R
\name{format_tracklog}
\alias{format_tracklog}
\title{Format Track log data}
\usage{
format_tracklog(
  data_dir = NULL,
  transect_layout,
  buffer = 10,
  write_output = TRUE,
  out_dir = fs::path(PEMprepr::read_fid()$dir_20105020_clean_field_data$path_rel),
  out_name = "s1_tracks_raw.gpkg"
)
}
\arguments{
\item{data_dir}{text string with location of raw files in shp or gpk format}

\item{transect_layout}{sf object with simplified transect layout}

\item{buffer}{numeric value for buffer distance around transect layout}

\item{write_output}{should the sf object be written to disk?
If \code{TRUE} (default), will write to \code{out_dir} under the appropriate resolution subfolder.}

\item{out_dir}{A character string of path which points to output location. A default
location and name are applied in line with standard workflow.}

\item{out_name}{A character string of the output file name. Default is \code{s1_tracks.gpkg}}
}
\value{
A\code{sf} multiline vector with the formatted track data
}
\description{
Standardize tracks collected during the field sampling process
}
\examples{
\dontrun{
clean_tracks <- format_tracklog(inputfolder, transect_layout, buffer = 10)
}
}
