% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_transectlayout.R
\name{generate_transectlayout}
\alias{generate_transectlayout}
\title{Generate Transect layout}
\usage{
generate_transectlayout(input_path)
}
\arguments{
\item{input_path}{filepath location where final sample plan or plans are stored}
}
\value{
sf vector of lines
}
\description{
Generate Transect layout
}
\examples{
create_transectlayout("inputfolder")

}
\author{
Genevieve Perkins
}
