% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEMsamplr-package.R
\docType{package}
\name{PEMsamplr-package}
\alias{PEMsamplr}
\alias{PEMsamplr-package}
\title{PEMsamplr: This sub-package generates a sample plan for predictive ecosystem mapping (PEM) workflow}
\description{
PEMsamplr is a subpackage of "PEMr" and forms part of the workflow to develop Predictive Ecosystem Maps (PEM). Specifically this package builds on data prepared using the PEMprepr package and generates a field sampling plan to collect data for modeling.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ninoxconsulting/PEMsamplr/issues}
}

}
\author{
\strong{Maintainer}: Genevieve Perkins \email{gperkins@ninoxconsulting.ca} (\href{https://orcid.org/0000-0003-2663-5904}{ORCID})

Authors:
\itemize{
  \item Will McKenzie \email{Will.MacKenzie@gov.bc.ca}
  \item Kiri Daust \email{kiri.daust@gov.bc.ca}
  \item Colin Chisholm \email{Colin.Chisholm@unbc.ca}
  \item Matthew Coghill \email{matthewcoghill@live.ca}
  \item Claire Armour \email{claire.armour@gov.bc.ca}
  \item Andy Teucher \email{andy@andyteucher.ca} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
