% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cost_layer.R
\name{prep_cost_layer}
\alias{prep_cost_layer}
\title{prepare_cost_layers}
\usage{
prep_cost_layer(
  vec_dir = PEMprepr::read_fid()$dir_1010_vector$path_abs,
  sampleplan_dir = PEMprepr::read_fid()$dir_201010_inputs$path_abs,
  heli = FALSE,
  out_dir = PEMprepr::read_fid()$dir_201010_inputs$path_abs,
  write_output = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{vec_dir}{directory where clean vector layers are stored}

\item{sampleplan_dir}{directory where sample plan inputs are stored}

\item{heli}{TRUE or FALSE to indicate if road or heli design}

\item{out_dir}{A \code{character} or path where cost layer is written to.}

\item{write_output}{A \code{logical}if the sf object be written to disk?
If \code{TRUE} (default), will write to \code{out_dir} under the appropriate resolution subfolder.}

\item{overwrite}{A \code{logical} if the output file already exists, should it be overwritten?}
}
\value{
\strong{SpatRast} cost surface to be used in cost layer creation
}
\description{
Use roads speed and slope to calculate initial cost of travel to be used in cost layer
}
\examples{
\dontrun{
prep_cost_layers(vec_dir,  sampleplan_dir, heli, write_output, out_dir)
}
}
