% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cost_layers.R
\name{prep_cost_layers}
\alias{prep_cost_layers}
\title{prepare_cost_layers}
\usage{
prep_cost_layers(vec_dir, dem)
}
\arguments{
\item{vec_dir}{directory where clean vector layers are stored}

\item{dem}{digital terrain model .SpatRast}
}
\value{
\strong{SpatRast} cost surface to be used in cost layer creation
}
\description{
Use roads speed and slope to calculate initial cost of travel to be used in cost layer
}
\examples{
prep_cost_layers(vec_dir, dem)


}
