% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cost_layers_gdistance.R
\name{prep_cost_layers_gdistance}
\alias{prep_cost_layers_gdistance}
\title{prepare_cost_layers}
\usage{
prep_cost_layers_gdistance(vec_dir, dem, heli = FALSE)
}
\arguments{
\item{vec_dir}{directory where clean vector layers are stored}

\item{dem}{digital terrain model .SpatRast}

\item{heli}{TRUE or FALSE to indicate if road or heli design}
}
\value{
\strong{SpatRast} cost surface to be used in cost layer creation
}
\description{
Use roads speed and slope to calculate initial cost of travel to be used in cost layer
}
\examples{
prep_cost_layers_gdistance(vec_dir, dem)


}
