% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tps_att.R
\name{prep_tps_att}
\alias{prep_tps_att}
\title{Prepare training pts attribute}
\usage{
prep_tps_att(
  clean_dir = fid$trainpts_maps[2],
  res = 5,
  cov_dir = fid$cov_dir_1020[2]
)
}
\arguments{
\item{clean_dir}{location of clean data point file}

\item{cov_dir}{location of output file}
}
\value{
datatable with training points
}
\description{
Prepares training points by adding neighbours, attributing values
}
\examples{
prep_tps_att (clean_dir = fid$trainpts_maps[2], res = 5,cov_dir = fid$cov_dir_1020[2])
}
