% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tps_bgc.R
\name{prep_tps_bgc}
\alias{prep_tps_bgc}
\title{Prepare training pts bgcs}
\usage{
prep_tps_bgc(
  allpts,
  mapkey,
  field = "Full",
  bec_path = file.path(fid$shape_dir_1010[2], "bec.gpkg"),
  min_no = 10
)
}
\arguments{
\item{allpts}{attributed points}

\item{mapkey}{mapkey}

\item{field}{string field withtin the mapkey to model, default = "FULL"}

\item{bec_path}{string filepath to bec.gpkg}

\item{min_no}{minimum number of points for a map unit, default = 10}
}
\value{
datatable points
}
\description{
Prepare training pts bgcs
}
\examples{
prep_tps_bgc(allpts, mapkey, field = "Full",bec_path = file.path(fid$shape_dir_1010[2], "bec.gpkg"),min_no = 10)
}
