% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_lowcost_clhs.R
\name{select_lowcost_clhs}
\alias{select_lowcost_clhs}
\title{Select lowest cost clhs option}
\usage{
select_lowcost_clhs(clhs_outpath)
}
\arguments{
\item{clhs_outpath}{filepath in which the clhs multiple outputs are saved}
}
\value{
plot to show the cost outputs per generated sample plan repeat
}
\description{
Select lowest cost clhs option
}
\examples{
select_lowcost_clhs(fid$samplingplan_clhs[2])
}
