% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vrp.R
\name{setup_python}
\alias{setup_python}
\title{Vehicle Routing Problem}
\usage{
setup_python()
}
\arguments{
\item{transition}{gdistance transition layer}

\item{clhs_points}{sample locations (sf)}

\item{start_pos}{start position (sf)}

\item{number_days}{number of days to sample for}
}
\description{
Vehicle Routing Problem
}
\examples{
#'
}
\author{
Kiri Daust
}
