% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_layer.R
\docType{data}
\name{x}
\alias{x}
\title{create cost layer}
\format{
An object of class \code{SpatRaster} of dimension 284 x 284 x 1.
}
\usage{
x
}
\arguments{
\item{x}{**SpatRast or RasterLayer (raster package), output from prepare_cost_layer ()}

\item{start}{**SpatVec of sf object with Start Locations for the cost layer.}
}
\value{
RasterLayer with Cost
}
\description{
create cost layer
}
\examples{
create_cost(costprep, start)
}
\keyword{datasets}
