% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_points.R
\name{attribute_points}
\alias{attribute_points}
\title{Add covariate predictor data to training point file}
\usage{
attribute_points(dat_pts, cov_dir)
}
\arguments{
\item{dat_pts}{a file gpkg of training points}

\item{cov_dir}{folder containing covariate files}
}
\value{
an sf object
}
\description{
Add covariate predictor data to training point file
}
\examples{
tpoints_ne <- attribute_points(dat_pts, cov_dir)
}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
