% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site_transects.R
\name{build_site_transects}
\alias{build_site_transects}
\title{Build site transects
Build the associates triangles, buffered triangles and selected paired triangle to sample, based on lowest cost.}
\usage{
build_site_transects(
  sample_points,
  cost,
  mask_poly,
  centroid_distance = 400,
  out_dir,
  outname = "s1_sampling.gpkg"
)
}
\arguments{
\item{sample_points}{a \code{sf} spatial object of clhs points}

\item{cost}{A \code{SpatRast} with cost layer generate for sample plan}

\item{mask_poly}{A\code{sf} spatial object of mask for specific bgc}

\item{centroid_distance}{\code{Numeric} value at which the triangles are placed apart,
default is 400 based on albers crs meters}

\item{out_dir}{A \code{character} or path where output sample plan is written to.}

\item{outname}{A \code{character} name for output file. Default is \code{s1_sampling.pgkg}.}
}
\value{
an \code{sf} geopackage with multiple layers
}
\description{
Build site transects
Build the associates triangles, buffered triangles and selected paired triangle to sample, based on lowest cost.
}
\examples{
\dontrun{
build_site_transects(sample_points, cost, mask_poly, centroid_distance = 400,
out_dir, outname = "s1_sampling.gpkg")
}
}
