% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bgc_cost.R
\name{check_bgc_cost}
\alias{check_bgc_cost}
\title{Check BGC sample cost}
\usage{
check_bgc_cost(bgc, binned_landscape, cost)
}
\arguments{
\item{bgc}{Biogeoclimatic classification vector (sf polygon/multipolygon)}

\item{binned_landscape}{Raster (SpatRast) with binned landscapes, output of create_binned_landscape() function}

\item{cost}{Raster layer (SpatRast) with cost layer}
}
\value{
data.frame
}
\description{
Assess the costs of sampling areas for each BGC to help assess how well the cost layer describes to study area
}
\examples{
check_bgc_cost(bgc, binned_landscape, cost)

}
