% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bgc_landscapes.R
\name{check_bgc_landscapes}
\alias{check_bgc_landscapes}
\title{Check landscapes per BGC}
\usage{
check_bgc_landscapes(bec, binned_landscape)
}
\arguments{
\item{bec}{vector spatial data (geopackage) of BEC within the study area}

\item{binned_landscape}{spatrast generated using the create_binned_landscape function}
}
\value{
\strong{data frame}
}
\description{
Assess how well the binned landscape are described for each of the BEC units of
interest
}
\examples{

check_bgc_landscapes(bec, binned_landscape)
}
