% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_lines_pts.R
\name{convert_lines_pts}
\alias{convert_lines_pts}
\title{Convert Line segments to training points}
\usage{
convert_lines_pts(processed_transects, trast)
}
\arguments{
\item{processed_transects}{sf object of all attributed line segments}

\item{trast}{raster template with the resolution}
}
\value{
sf object points based on raster
}
\description{
Convert Line segments to training points
}
\examples{
convert(lines, trast)

}
\author{
Genevieve Perkins
}
