% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_lines_pts.R
\name{convert_lines_pts}
\alias{convert_lines_pts}
\title{Convert Line segments or points to rasterized training points}
\usage{
convert_lines_pts(
  processed_lines = fs::path(PEMprepr::read_fid()$dir_20105020_clean_field_data$path_rel,
    "proc_s1_transects5.gpkg"),
  trast,
  buffer = 2.5,
  write_output = TRUE,
  out_dir = PEMprepr::read_fid()$dir_20105020_clean_field_data$path_rel,
  out_name = "allpoints.gpkg"
)
}
\arguments{
\item{processed_lines}{A \code{sf} object of all attributed line segments or points}

\item{trast}{A \code{SpatRast} or path to file with resolution matching the modelling resolution.}

\item{buffer}{A \code{numeric}to determine the extend in meters the line segments will be expanded to.
Default is half of the raster template resolution 2.5m for a 5m resolution raster template}

\item{write_output}{A \code{logical}if the sf object be written to disk?
If \code{TRUE} (default), will write to \code{out_dir} under the appropriate resolution subfolder.}

\item{out_dir}{A character string of path which points to output location. A default
location and name are applied in line with standard workflow.}

\item{out_name}{A character string of the output file name. Default is \code{allpoints.gpkg}}
}
\value{
sf object points based on raster reolution
}
\description{
Convert Line segments or points to rasterized training points
}
\examples{
\dontrun{
clean_tracks <- convert_to_pts(
  processed_lines = fs::path(PEMprepr::read_fid()$dir_20105020_clean_field_data$path_rel,
  "proc_s1_transects5.gpkg"),
  trast = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel, "5m", "template.tif"),
  buffer = 2.5,
  write_output = TRUE,
  out_dir = PEMprepr::read_fid()$dir_20105020_clean_field_data$path_rel,
  out_name = "allpoints.gpkg")
}
}
