% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bgc_mask.R
\name{create_bgc_mask}
\alias{create_bgc_mask}
\title{Create cost no sample area}
\usage{
create_bgc_mask(vec_dir, cost_masked, out_dir)
}
\arguments{
\item{vec_dir}{text string with folder location of base vector layers}

\item{cost_masked}{SpatRast of cost layer with high cost penalty applied}

\item{out_dir}{text string with location of output vector layers}
}
\value{
\strong{SpatRast} A masked cost raster layer
}
\description{
Creates and applies a mask to permanently changed areas where sample is not be conducted.
This includes (lakes, and permanently changed landscape i.e roads)
}
\examples{
create_cost_nosample <- function(vec_dir, cost)

}
