% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bgc_mask.R
\name{create_bgc_mask}
\alias{create_bgc_mask}
\title{Create cost no sample areas per bec zone}
\usage{
create_bgc_mask(
  vec_dir = fs::path(PEMprepr::read_fid()$dir_1010_vector$path_abs),
  cost_masked,
  out_dir = fs::path(PEMprepr::read_fid()$dir_201010_inputs$path_abs),
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{vec_dir}{text string with folder location of base vector layers}

\item{cost_masked}{A \code{SpatRast} of cost layer with high cost penalty and
exclusions applied.}

\item{out_dir}{A \code{character} or path which points to input location
of . A default location and name are applied in line with standard workflow.}

\item{overwrite}{logical. If \code{TRUE}, \code{filename} is overwritten}

\item{...}{Additional options passed on to \code{...} in \code{\link[terra:writeRaster]{terra::writeRaster()}}}
}
\value{
path to the output directory where files are written (invisibly).
}
\description{
Creates and applies a mask of the bec zone to the final cost layer
}
\examples{
\dontrun{
create_bgc_mask <- function(vec_dir, cost_final, out_dir)
}

}
