% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_binned_landscape.R
\name{create_binned_landscape}
\alias{create_binned_landscape}
\title{Create binned landscape}
\usage{
create_binned_landscape(
  input,
  covariates = c("dah_LS", "landscape_LS", "mrvbf_LS")
)
}
\arguments{
\item{input}{text string with either folder location or individual files}

\item{covariates}{text string with file names (default to dah, landscape, mrvbf)}
}
\value{
\strong{SpatRaster}
}
\description{
Create a landscape with binned variables to assess the landscape environmental space.
}
\details{
Can use either folder or individual files to assess landscape covariate space
}
\examples{

create_binned_landscape(input,covariates = c("dah_LS", "landscape_LS","mrvbf_LS"))
}
