% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clhs.R
\name{create_clhs}
\alias{create_clhs}
\title{Create CLHS sample plan}
\usage{
create_clhs(
  all_cov,
  num_slices,
  to_include = NULL,
  n_points = 5,
  min_dist = 900,
  num_sample = 5e+06
)
}
\arguments{
\item{all_cov}{Terra rast of LS covariates and cost layer}

\item{num_slices}{Number of slices}

\item{to_include}{sf dataframe of already sample points to include in plan}

\item{n_points}{Number of points per slice}

\item{min_dist}{Minimum distance between points. Note that this is not guaranteed, especially with lots of points in a small area}

\item{num_sample}{Number of samples to run CLHS on}
}
\value{
sf dataframe with sampled points and labels for slice number and point number
}
\description{
This function takes a terra stack of covariates (and cost), and uses conditional latin hypercube sampling to create the specified number of points in slices.
}
\author{
Kiri Daust
}
