% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_layer.R
\name{create_cost_layer}
\alias{create_cost_layer}
\title{create cost layer}
\usage{
create_cost_layer(x, start)
}
\arguments{
\item{x}{**SpatRast or RasterLayer (raster package), output from prepare_cost_layer ()}

\item{start}{**SpatVec of sf object with Start Locations for the cost layer.}
}
\value{
RasterLayer with Cost
}
\description{
create cost layer
}
\examples{
create_cost(costprep, start)
}
