% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_penalty.R
\name{create_cost_penalty}
\alias{create_cost_penalty}
\title{Create cost penalty layer}
\usage{
create_cost_penalty(
  vec_dir,
  dem,
  cost,
  costval = 3000,
  vri_cost = 2500,
  calc_by_qq = TRUE
)
}
\arguments{
\item{vec_dir}{text string with folder location of base vector layers}

\item{dem}{SpatRast or Raster of cost layer}

\item{cost}{SpatRast or Raster of cost layer}

\item{costval}{Numeric value of assigned high cost (ie default is 3000)}

\item{vri_cost}{Numeric value of assigned secondary high cost (i.e default is 2500)}

\item{calc_by_qq}{TRUE/FALSE if cost is assigned by distribution rather than specified in cost_val anc vri_cost}
}
\value{
SpatRast
}
\description{
Create cost penalty layer
}
\examples{
create_cost_penalty(vec_dir, dem, cost, costval = 3000)

}
