% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_penalty.R
\name{create_cost_penalty}
\alias{create_cost_penalty}
\title{Create cost penalty layer}
\usage{
create_cost_penalty(vec_dir, dem, cost, costval = 3000)
}
\arguments{
\item{vec_dir}{text string with folder location of base vector layers}

\item{dem}{SpatRast or Raster of cost layer}

\item{cost}{SpatRast or Raster of cost layer}

\item{costval}{Numeric value of assigned high cost}
}
\value{
SpatRast
}
\description{
Create cost penalty layer
}
\examples{
create_cost_penalty(vec_dir, dem, cost, costval = 3000)

}
