% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_penalty.R
\name{create_cost_penalty}
\alias{create_cost_penalty}
\title{Create a cost penalty layer for sample plan development}
\usage{
create_cost_penalty(
  vec_dir,
  dem,
  cost,
  costval = 3000,
  vri_cost = 2500,
  calc_by_qq = TRUE,
  out_dir = fs::path(PEMprepr::read_fid()$dir_201010_inputs$path_abs),
  write_output = TRUE
)
}
\arguments{
\item{vec_dir}{A \code{character} string or path which points to base vector
layers created using create_base_vectors(). A default location and name
are applied in line with standard workflow.}

\item{dem}{A \code{SpatRast} or path to the landscape scale DEM data. A default
location and name is applied in line with standard workflow.}

\item{cost}{A \code{SpatRast} or path to cost layer created using the
create_cost_layer() function. A default location and name is applied in
line with standard workflow.}

\item{costval}{A \code{numeric} value representing a high cost value. Default is
3000. These values are only used when calc_by_qq is FALSE.}

\item{vri_cost}{A \code{numeric} value representing the cost to be applied to VRI
age class 1 and 2 cost value. Default is 2500.These values are only used
when calc_by_qq is FALSE.}

\item{calc_by_qq}{should the cost values be calculated based on distribution
of cost values rather than specified in costval and vri_cost. The costs
assigned by distribution are 0.65 for vri_cost, 0.70 for costval and 0.9
for max cost, applied to the slope greater than 45\% degrees.}

\item{out_dir}{A \code{character} or path which points to input location
of . A default location and name are applied in line with standard workflow.}

\item{write_output}{A \code{logical} should the cost_penalty spatRaster be
written to disk? If \code{TRUE} (default), will write to \code{out_dir}.}
}
\value{
a \code{spatRaster} representing costs with cost penality values applied
}
\description{
Uses cost layer and applies a cost layers to the following criteria:
cutblocks (within the last 20 years), vri (age class 1 & 2), vri age class 3,
vri (deciduous leading), private land, fire intensity, fire boundaries,
transmission lines and steep slopes.
}
\examples{
\dontrun{
sampleplan_dir <- PEMprepr::read_fid()$dir_201010_inputs$path_abs
acost <- terra::rast(fs::path(sampleplan_dir, "acost.tif" ))
terra::crs(acost) <- terra::crs(bec)
names(acost) <- "cost"
vec_dir = fs::path(PEMprepr::read_fid()$dir_1010_vector$path_abs)
dem = terra::rast(fs::path(PEMprepr::read_fid()$dir_201010_inputs$path_abs, "25m", "dem.tif"))
cost =  acost
costval = 3000
vri_cost = 2500
calc_by_qq = TRUE

cost_penalty <- create_cost_penalty(vec_dir = vec_dir,
                                   cost = acost,
                                   dem = dem,
                                   costval = 3000,
                                   vri_cost = 2500,
                                   calc_by_qq = TRUE,
                                   write_output = FALSE)


}


}
