% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampleplan.R
\name{create_sampleplan}
\alias{create_sampleplan}
\title{Create sample plan including spatial and tracking files}
\usage{
create_sampleplan(
  clhs_set,
  clhs_dir = PEMprepr::read_fid()$dir_20103010_clhs$path_abs,
  mask_dir = PEMprepr::read_fid()$dir_201020_masks$path_abs,
  cost_dir = PEMprepr::read_fid()$dir_201010_inputs$path_abs,
  out_dir = PEMprepr::read_fid()$dir_20103020_review$path_rel
)
}
\arguments{
\item{clhs_set}{A \code{character} containing the names of clhs files to be converted
to full sample plan. One or more BEC units can be selected for a sampleplan}

\item{clhs_dir}{A path to the location of files specified in \code{clhs_set} above.
Default location is based on standard workflow.}

\item{mask_dir}{A path to the location of masked bec spatial files for the BEC
units specified by the clhs_set. Default location is based on standard workflow.}

\item{cost_dir}{A path to the location of final cost output \code{spatRast}.
Default location is based on standard workflow.}

\item{out_dir}{A path to the location in which the output sample plan will be
saved. Default location is based on standard workflow.}
}
\value{
path to the output directory where files are written (invisibly).
}
\description{
This function used the selected clhs outputs and generates a full sampleplan
spatial data in geopackage output along with a csv tracking sheet per BEC
unit
}
\examples{
\dontrun{
generate_sampleplan (
  clhs_set <- c("ICHmc2_clhs_sample_3.gpkg", "ICHmc1_clhs_sample_3.gpkg"),
  clhs_dir = PEMprepr::read_fid()$dir_20103010_clhs$path_abs,
  mask_dir = PEMprepr::read_fid()$dir_201020_masks$path_abs,
  cost_dir = PEMprepr::read_fid()$dir_201010_inputs$path_abs,
  out_dir = PEMprepr::read_fid()$dir_20103020_review$path_rel)
}
}
