% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_samplr_covariates.R
\name{create_samplr_covariates}
\alias{create_samplr_covariates}
\title{Generate landscape level covariate rasters for cLHS sample plan}
\usage{
create_samplr_covariates(
  dtm = dtm,
  rtemplate = rtemplate,
  SAGApath = "",
  output = NULL,
  covariates = c("mrvbf", "dah", "landform"),
  sieve_size = 10,
  dah_threshold = 0.2,
  saga_param = list(T_SLOPE = 64, TPCTL_V = 6, T_PCTL_R = 2, P_SLOPE = 4, P_PCTL = 3,
    UPDATE = 1, CLASSIFY = 1, MAX_RES = 100)
)
}
\arguments{
\item{dtm}{is a 25m dtm raster object ideally cropped to the AOI watershed boundaries}

\item{rtemplate}{template of 25m raster to match final output to}

\item{SAGApath}{Is the location of SAGA on your system.  On linux systems with SAGA GIS installed Use \code{SAGApath = ""}}

\item{output}{Location of where rasters will be saved.}

\item{covariates}{character vector of covariates to create. Can be any of c("mrvbf", "dah", "landform")}

\item{sieve_size}{Remove isolated clusters of below the threshold number of cells}

\item{dah_threshold}{dah threshold to classify the dah layer into the three respective classes}

\item{saga_param}{named list. Parameters that are fed to SAGA MRVBF function. IF supplied it must contain all parameters specified in the SAGA command.}
}
\description{
Generates base landscape covariates from a 25m TRIM DEM using SAGA GIS and converts to classes for use in the sample plan cLHS
}
\details{
This script has been tested with SAGA 8.4 on Windows
Depending on your system the path to \code{saga_cmd} may need to be specified.
}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
