% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vehicle_vrp.R
\name{create_vrp}
\alias{create_vrp}
\title{Vehicle Routing Problem}
\usage{
create_vrp(pnts, start, tr1)
}
\arguments{
\item{pnts}{sp object with clhs generated points}

\item{start}{start location}

\item{tr1}{transition layer (.rds) format created in cost layer step}
}
\description{
Vehicle Routing Problem
}
