% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vehicle_vrp.R
\name{create_vrp}
\alias{create_vrp}
\title{Vehicle Routing Problem}
\usage{
create_vrp(transition, clhs_points, start_pos, number_days)
}
\arguments{
\item{transition}{gdistance transition layer}

\item{clhs_points}{sample locations (sf)}

\item{start_pos}{start position (sf)}

\item{number_days}{number of days to sample for}
}
\description{
Vehicle Routing Problem
}
\examples{
#'
}
\author{
Kiri Daust
}
