% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fielddata.R
\name{format_fielddata}
\alias{format_fielddata}
\title{Format field data}
\usage{
format_fielddata(datafolder, transect_layout_buf)
}
\arguments{
\item{datafolder}{text string with location of raw files in shp or gpk format}

\item{transect_layout_buf}{sf spatial object with transect buffered data}
}
\value{
sf point data
}
\description{
Standardizes format from raw field data to standard point data
}
\examples{

clean_pts <- format_fielddata(inputfolder, transect_layout)
}
\author{
Genevieve Perkins
}
