% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tracklog.R
\name{format_tracklog}
\alias{format_tracklog}
\title{Format Track log data}
\usage{
format_tracklog(datafolder, transect_layout_buf)
}
\arguments{
\item{datafolder}{text string with location of raw files in shp or gpk format}

\item{transect_layout_buf}{sf spatial object with transect buffered data}
}
\value{
sf multiline vector
}
\description{
Standardize tracks collected during the field sampling process
}
\examples{
clean_tracks <- format_tracklog(inputfolder, transect_layout)
}
