% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tracklog.R
\name{format_tracklog}
\alias{format_tracklog}
\title{Format Track log data}
\usage{
format_tracklog(data_dir = NULL, transect_layout, buffer = 10)
}
\arguments{
\item{data_dir}{text string with location of raw files in shp or gpk format}

\item{transect_layout}{sf object with simplified transect layout}

\item{buffer}{numeric value for buffer distance around transect layout}
}
\value{
A\code{sf} multiline vector
}
\description{
Standardize tracks collected during the field sampling process
}
\examples{
\dontrun{
clean_tracks <- format_tracklog(inputfolder, transect_layout, buffer = 10)
}

}
