% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mapkey.R
\name{generate_mapkey}
\alias{generate_mapkey}
\title{Generate a mapkey from provincial BEC unit list and flag unmatched field data}
\usage{
generate_mapkey(
  data_pts,
  key = utils::read.csv(fs::path_package("PEMsamplr", "extdata/mapkey_all_draft.csv")),
  write_output = FALSE,
  out_dir = fs::path(PEMprepr::read_fid()$dir_3010_inputs$path_rel),
  out_name = "mapunitkey.csv"
)
}
\arguments{
\item{data_pts}{A \code{sf}object with standardised attribute fields}

\item{key}{A \code{data.frame} with the provincial BEC unit list. This is an internal dataset}

\item{write_output}{A \code{logical}if the map key should be be written to disk?
If \code{TRUE} (default), will write to \code{out_dir} under the appropriate resolution subfolder.}

\item{out_dir}{A character string of path which points to output location. A default
location and name are applied in line with standard workflow.}

\item{out_name}{A character string of the output file name. Default is \code{mapunitkey.csv}}
}
\value{
a tibble with unique mapkey for the study area along with unmatched fieldcalls to manually review
}
\description{
Generate a mapkey from provincial BEC unit list and flag unmatched field data
}
\examples{
\dontrun{
key <- generate_mapkey(
  data_pts = fs::path(
    PEMprepr::read_fid()$dir_20105020_clean_field_data$path_rel,
    "s1_points_raw.gpkg"
  ),
  key = utils::read.csv(fs::path_package("PEMsamplr", "extdata/mapkey_all_draft.csv")),
  write_output = FALSE,
  out_dir = fs::path(PEMprepr::read_fid()$dir_3010_inputs$path_rel),
  out_name = "mapunitkey.csv"
)
}
}
