% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lines.R
\name{make_lines}
\alias{make_lines}
\title{make_lines()}
\usage{
make_lines(
  GPSPoints = NA,
  GPSTracks = NA,
  Transects,
  method = "pts2lines",
  sortby = "none",
  tBuffer = 20,
  PROJ = 3005
)
}
\arguments{
\item{GPSPoints}{is a \emph{simple features} POINT object.
Minimum attributes in this file include fields named \emph{'id'} and \emph{'time'}.
Waypoints will be sorted by these two entries.}

\item{Transects}{is the original planned transect simple feature.
The \emph{'id'} attribute in this feature will be used to group points by their transect id.
This transect id will be named 'TID' in the output data.}

\item{method}{Multiple methods to be made available:
\itemize{
\item \emph{pts2lines} takes a sf POINT object and converts it to lines.
_This is currently the only function available
\item \emph{tracklog} uses the tracklog and the sample points to generate the lines. \emph{Not implemented yet}.
}}

\item{sortby}{Field in the points data to sort the points by.  Defaults to  \emph{"none"} (i.e. assumes no sorting of data)}

\item{tBuffer}{is optional with a default of 20m.  This is the buffer distance to place around the transect.  Waypoints outside this distance will not be considered.}

\item{PROJ}{is an optional the PROJ projection code with a default of BC Albers (3005).  Data imported will be transfored to this projection and final data will be exported in this projection.}
}
\description{
Converts GPS data/ waypoints to line features.  Attribute data is maintained.
Currently, this function generates points to transect by converting the \emph{POINT}
feature data to LINES sf features.
A placeholder has been made to generate transect lines via the transect method created by M. Coghill.
}
\examples{
## Convert GPS waypoints to line features (i.e. transect data)
transects <- convert_pts2lines(gpsData, PlannedTransects)
}
\keyword{convertion}
\keyword{lines,}
\keyword{points,}
