% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lines.R
\name{make_lines}
\alias{make_lines}
\title{Convert point or tracklog to line segments}
\usage{
make_lines(
  points = NA,
  transect_layout,
  method = "pts2lines",
  buffer = 20,
  write_output = TRUE,
  out_dir = fs::path(PEMprepr::read_fid()$dir_20105020_clean_field_data$path_rel),
  out_name = "proc_s1_transects.gpkg"
)
}
\arguments{
\item{points}{A \code{sf}object with standardised attribute fields}

\item{transect_layout}{A \code{sf} object with simplified transect layout.}

\item{method}{Multiple methods to be made available:
\itemize{
\item \emph{pts2lines} takes a sf POINT object and converts it to lines.
_This is currently the only function available
\item \emph{tracklog} uses the tracklog and the sample points to generate the lines. \emph{Not implemented yet}.
}}

\item{buffer}{A numeric value for buffer distance around transect layout a default of 20m.
This is the buffer distance to place around the transect.
Waypoints outside this distance will not be considered.}

\item{write_output}{should the sf object be written to disk?
If \code{TRUE} (default), will write to \code{out_dir} under the appropriate resolution subfolder.}

\item{out_dir}{A character string of path which points to output location. A default
location and name are applied in line with standard workflow.}

\item{out_name}{A character string of the output file name. Default is \code{proc_s1_transects.gpkg}}
}
\description{
Converts standardized points to line segments.
Currently, this function generates points to transect by converting the \emph{POINT}
feature data to LINES sf features.
A placeholder has been made to generate transect lines via the transect method created by M. Coghill.
}
\examples{
\dontrun{
processed_lines <- make_lines(points, transect_layout, buffer = 20, method = "pts2lines")
}
}
