% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clhs_cost.R
\name{plot_clhs_cost}
\alias{plot_clhs_cost}
\title{Plot the cost of clhs sample plans}
\usage{
plot_clhs_cost(clhs_dir = PEMprepr::read_fid()$dir_20102010_clhs$path_rel)
}
\arguments{
\item{clhs_dir}{A path or directory where the multiple clhs sample plans are saved.
Default location is based on standard workflow.}
}
\value{
ggplot2 plot
}
\description{
Plot the cost of clhs sample plans
}
\examples{
\dontrun{
check_clhs_cost(PEMprepr::read_fid()$dir_20102010_clhs$path_abs)
}
}
