% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tps_covs.R
\name{prep_tps_covs}
\alias{prep_tps_covs}
\title{Prepare training pts bgcs}
\usage{
prep_tps_covs(tpts, mcols, outdir = fid$model_inputs0310[2])
}
\arguments{
\item{tpts}{dataframe containing attributed and bgc selected points}

\item{mcols}{string list of cov name}

\item{outdir}{location of output files for model, default selected based on strandard stucture}
}
\value{
dataframe of final training pts
}
\description{
Prepare training pts bgcs
}
\examples{
prep_tps_covs(tpts, mcols, outdir = fid$model_inputs0310[2])
}
