% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_transectlayout.R
\name{simplify_transectlayout}
\alias{simplify_transectlayout}
\title{Simplify sample plan layout for post processing}
\usage{
simplify_transectlayout(
  input_path = fs::path(PEMprepr::read_fid()$dir_20104020_transect$path_rel),
  out_dir = fs::path(PEMprepr::read_fid()$dir_20104020_transect$path_rel),
  writeout = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{input_path}{A \code{character} string or path where sample plan layout geopackage is stored.
Default location is based on standard workflow. Note this function is set up to work with
multiple files so ensure sample plan is only .gpkg in folder.}

\item{out_dir}{A path to the location in which the simplified transect layout will be
saved. Default location is based on standard workflow.}

\item{writeout}{should the simplifeid transect layout sf object be written to disk?
If \code{TRUE} (default), will write to \code{out_dir}. Default location is based
on standard workflow.}

\item{overwrite}{a \code{logical} to determine if the output file be overwritten
if it already exists? Only used when \code{writeout = TRUE}. Default is \code{FALSE}.}
}
\value{
sf object of simplified transect layout
}
\description{
Simplify sample plan layout for post processing
}
\examples{
\dontrun{
t1 <- simplify_transectlayout(
    input_dir = fs::path(PEMprepr::read_fid()$dir_20104020_transect$path_abs),
    out_dir = fs::path(PEMprepr::read_fid()$dir_20104020_transect$path_abs),
    writeout = TRUE,
    overwrite = FALSE)
}
}
